Improvements

+++ src/impl/x86_64/string.c
@@ -32,10 +32,10 @@ int strlen(const char* str) {
     return len;
 }
 
-void* memcpy(void* dest, const void* src, int count) {
+void* memcpy(void* dest, const void* src, size_t count) {
     char* dst = (char*)dest;
     const char* source = (const char*)src;
-    for (int i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         dst[i] = source[i];
     }
     return dest;
